Update Mocha & Istanbul
We were on a really old version of Mocha, and standard practice for JS
code coverage is to use `nyc` now. This switches our code coverage tool
to `nyc`, and gets us on the latest version of Mocha.

Our code coverage did not appear to be working before either,
now it does.

Additionally, Mocha now supports `.mocharc.js` files which let us have
inline comments for configuration. `.mocha.opts` is deprecated, so I
have switched our Mocha configuration to the new format.

If we had configuration that was equal to the default, I have removed it