Skip coverage for untestable code (#83)
This way we can get to 100% code coverage. 100% coverage is great
because it gives a clear indicator when there has been a coverage
regression, and it makes explicit which code you are choosing not to
test (through inline coverage disable comments) rather than implicit
(in lower percentages of coverage in the report).